fixed issue with SR not displaying correctly

+++ src/GUI/EmulatorUI.java
@@ -26,7 +26,7 @@ public abstract class EmulatorUI  { // if there is no non-abstract methods, turn
 	
 	public abstract void updateAddressRegisterDisplay(int register, int data);
 	
-	public abstract void updateStatusRegister(int value);
+	public abstract void updateStatusRegister(short value);
 	
 
 }
